Refactor ClientService. Removi a Service e reparti em UseCases. Testado no postman ok
