+++ src/main/java/com/tech/challenge/tech_challenge/core/domain/useCases/CreateClientUseCase.java
@@ -1,0 +1,18 @@
+package com.tech.challenge.tech_challenge.core.domain.useCases;
+
+import com.tech.challenge.tech_challenge.adapters.driven.infra.repositories.ClientRepository;
+import com.tech.challenge.tech_challenge.core.application.exceptions.ValidationException;
+import com.tech.challenge.tech_challenge.core.domain.entities.Client;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+@Service
+public class CreateClientUseCase {
+
+    @Autowired
+    private ClientRepository clientRepository;
+    public Client execute(Client client) throws ValidationException {
+        client.validate();
+        return clientRepository.save(client);
+    }
+}
