+++ src/main/java/com/tech/challenge/tech_challenge/core/domain/useCases/FindClientsUseCase.java
@@ -1,0 +1,19 @@
+package com.tech.challenge.tech_challenge.core.domain.useCases;
+
+import com.tech.challenge.tech_challenge.adapters.driven.infra.repositories.ClientRepository;
+import com.tech.challenge.tech_challenge.core.domain.entities.Client;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class FindClientsUseCase {
+
+    @Autowired
+    private ClientRepository clientRepository;
+
+    public List<Client> execute(){
+        return clientRepository.findAll();
+    }
+}
