+++ src/test/java/com/tech/challenge/tech_challenge/adapters/driver/controllers/PaymentControllerTest.java
@@ -5,7 +5,6 @@ import com.tech.challenge.tech_challenge.core.application.exceptions.ResourceNot
 import com.tech.challenge.tech_challenge.core.application.exceptions.ValidationException;
 import com.tech.challenge.tech_challenge.core.application.message.MessageResponse;
 import com.tech.challenge.tech_challenge.core.domain.entities.*;
-import com.tech.challenge.tech_challenge.core.domain.services.ClientService;
 import com.tech.challenge.tech_challenge.core.domain.services.PaymentService;
 import com.tech.challenge.tech_challenge.core.domain.services.extended.OrderPaymentService;
 import org.junit.jupiter.api.BeforeEach;
