+++ src/test/java/com/tech/challenge/tech_challenge/adapters/driver/controllers/ProductControllerTest.java
@@ -6,7 +6,6 @@ import com.tech.challenge.tech_challenge.core.application.exceptions.UsedProduct
 import com.tech.challenge.tech_challenge.core.application.exceptions.ValidationException;
 import com.tech.challenge.tech_challenge.core.application.message.MessageResponse;
 import com.tech.challenge.tech_challenge.core.domain.entities.*;
-import com.tech.challenge.tech_challenge.core.domain.services.ClientService;
 import com.tech.challenge.tech_challenge.core.domain.services.ProductService;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
